auth_flask_app/requirements.txt
flask==2.2.2
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.4
requests==2.28.1
gevent==21.12.0
psycopg2-binary==2.9.3
flask_sqlalchemy==2.5.1
redis==4.3.4
gunicorn==20.1.0
flasgger==0.9.5
pydantic[dotenv]==1.9.2
Flask-Migrate==3.1.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
